% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vus.R
\name{ci_VUS}
\alias{ci_VUS}
\title{Confidence Intervals for Covariate-specific VUS}
\usage{
ci_VUS(x, ci.level = 0.95)
}
\arguments{
\item{x}{an object of class "VUS", a result of a call to \code{\link{VUS}}.}

\item{ci.level}{a confidence level to be used for constructing the confidence interval; default is 0.95.}
}
\value{
\code{ci_VUS} returns an object of class inheriting from "ci_VUS" class. An object of class "ci_VUS" is a list containing at least the following components:

\item{vus_ci_norm}{the normal-approximate confidence interval of covariate-specific VUS.}
\item{vus_ci_log}{the confidence interval of covariate-specific VUS after using logit-transformation.}
\item{vus_ci_prob}{the confidence interval of covariate-specific VUS after using probit-transformation.}
\item{ci.level}{confidence level is used.}
\item{x.val}{value(s) of covariate(s).}
\item{n_p}{total numbers of the regressors in the model.}
}
\description{
Computes confidence intervals for covariate-specific VUS obtained by function \code{\link{VUS}}.
}
\details{
A confidence interval of covariate-specific VUS also is given based on Normal-approximate. If the lower bound (or the upper bound) of the confidence interval is less than 0 (or greater than 1), it will be set as 0 or 1. Furthermore, logit and probit transformations are also applied in order to guarantee the confidence intervals are not outside (0,1).
}
\seealso{
\code{\link{VUS}}
}
