% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_3class.R
\docType{data}
\name{data_3class}
\alias{data_3class}
\title{A simulated data}
\format{
A data frame with 225 observations (from 30 clusters).
\describe{
  \item{\code{id_Clus}}{the id number of cluster.}
  \item{\code{Y}}{a vector containing test results.}
  \item{\code{D}}{a factor with 3 levels for the disease status, 1, 2, 3. The levels correspond to benign disease, early stage and late stage.}
  \item{\code{X1}}{a continuous covariate.}
  \item{\code{X2}}{a binary covariate.}
}
}
\usage{
data(data_3class)
}
\description{
A simulated data example with 30 clusters.
}
\keyword{data}
