% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_3class_bcx.R
\docType{data}
\name{data_3class_bcx}
\alias{data_3class_bcx}
\title{A simulated data}
\format{
A data frame with 582 observations (from 60 clusters).
\describe{
  \item{\code{id_Clus}}{the id number of cluster.}
  \item{\code{Y}}{a vector containing test results.}
  \item{\code{D}}{a factor with 3 levels for disease status, 1, 2, 3. The levels correspond to benign disease, early stage and late stage.}
  \item{\code{X}}{a continuous covariate.}
}
}
\usage{
data(data_3class_bcx)
}
\description{
A simulated data example with 60 clusters. This dataset is used in a example of analysis with Box-Cox transformation.
}
\keyword{data}
