% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lme2.R
\name{lme2}
\alias{lme2}
\title{Fitting the cluster-effect models for two-class of three-class settings.}
\usage{
lme2(
  name.test,
  name.class,
  name.covars,
  name.clust,
  data,
  apVar = TRUE,
  residual = FALSE,
  fitted = FALSE,
  randf = FALSE,
  ...
)
}
\arguments{
\item{name.test}{....}

\item{name.class}{....}

\item{name.covars}{...}

\item{name.clust}{...}

\item{data}{...}

\item{apVar}{a logical value ...}

\item{residual}{a logical value ...}

\item{fitted}{a logical value. Default = \code{FALSE}. If set to \code{TRUE}, ...}

\item{randf}{a logical value ...}

\item{...}{additional arguments for \code{lme()}.}
}
\value{
\code{lme2} returns an object of class inheriting from "lme2" class.

The function \code{\link{print.lme2}} can be used to print a summary of the results.

An object of class "lme2" is a list containing at least the following components:

\item{call}{the matched call.}
\item{est_para}{the estimate of VUS.}
\item{se_para}{the standard error, obtained by using asymptotic theory or bootstrap resampling method.}
\item{vcov_sand}{the matched call.}
\item{residual}{...}
\item{fitted}{...}
\item{randf}{...}
\item{n_coef}{...}
}
\description{
\code{lme2} the cluster-effect models for two-class or three-class setting based on the \code{lme()} routine from \code{nlme}-package.
}
\details{
.....
}
\examples{
## Example for two-class setting
data(data_2class)
head(data_2class)
out1 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1", "X2"), name.clust = "id_Clus",
             data = data_2class, apVar = FALSE)
print(out1)
out2 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1", "X2"), name.clust = "id_Clus",
             data = data_2class)
print(out2, digits = 3)

## Example for three-class setting
data(data_3class)
head(data_3class)
out3 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1", "X2"), name.clust = "id_Clus",
             data = data_3class, apVar = FALSE)
print(out3)
out4 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1", "X2"), name.clust = "id_Clus",
             data = data_3class)
print(out4)

}
