% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_3class.R
\name{optThres3}
\alias{optThres3}
\title{Estimate the covariate-specific optimal pair of thresholds for clustered data.}
\usage{
optThres3(
  method = c("GYI", "CtP", "MV"),
  out_lme2,
  x.val,
  apVar = TRUE,
  data,
  control = list()
)
}
\arguments{
\item{method}{The method to be used. See 'Details'.}

\item{out_lme2}{an object of class "lme2", a result of a call to \code{\link{lme2}}.}

\item{x.val}{specific value(s) of covariate(s) where the ROC surface are computed. In case non-covariate, no value is needed to specify. In case of one covariate, \code{x.val} should be a number. In case of \eqn{p} covariates (\eqn{p > 1}), \code{x.val} should be a vector containing \eqn{p} values of the covariates.}

\item{apVar}{a logical value. If set to TRUE, the variance-covariance matrix of covariate-specific optimal thresholds is computed.}

\item{data}{a data frame containing the variables in the model, which used for performing bootstrap procedure to estimate the variance-covariance matrix of covariate-specific optimal thresholds.}

\item{control}{a list of control parameters. See 'Details'}
}
\value{
\code{optThres3} returns an object of class inheriting from "optThres3" class, which is a list containing at least the following components:

\item{call}{the matched call.}
\item{method}{the methods were used to obtain the optima threholds.}
\item{thres3}{a vector or matrix containing the estimated optimal thresholds.}
\item{vcov.thres3}{a matrix or list of matrices containing the estimated variance-covariance matrices.}
\item{tcfs}{a vector or matrix containing the estimated TCFs at the optimal thresholds.}
\item{x.val}{value(s) of covariate(s).}
\item{n_p}{total numbers of the regressors in the model.}

Generic functions such as \code{print} and \code{plot} have methods to show the results.
}
\description{
\code{optThres3} estimates covariate-specific optimal pair of thresholds of a continuous diagnostic test in a clustered design when subjects can be diagnosed in three ordinal groups.
}
\details{
This function implements three estimation methods discussed in To et al. (2021) for estimating covariate-specific optimal pair of thresholds in a clustered design when subjects diagnosed in three ordinal groups. The estimators are based on the results of fitting the linear mixed-effect model on the diagnostic tests, which is done by using \code{\link{lme2}} with REML approach.

Before applying the estimation, a quick check for the monotone ordering assumption will be performed. That is, for given values of covariates, three predicted means of three diagnostic groups will be compared. If the assumption does not meet, the covariate-specific optimal pair of thresholds at the values of covariates will be not estimated.

Method \code{"GYI"} is Generalized Youden Index. This maximizes the sum of three covariate-specific True Class Fractions - TCFs (or total of correct classification rates) to obtain the optimal pair of thresholds. Method \code{"CtP"} is Closest to Pefection. By using this method, the optimal pair of thresholds is obtained by minimizing the distance, in the unit cube, between the point of three covariate-specific TCFs and the top corner (1, 1, 1). Method \code{"MV"} is Max Volume, which searches for thresholds that maximize the volume of a box under the covariate-specific ROC surface, and the volume is defined as the product of the three covariate-specific TCFs. The user can select more than one method.

The asymptotic variance-covariance matrix of the estimated covariate-specific optimal thresholds is approximated through the Delta method under the normal assumptions. If the Box-Cox transformation was applied for the linear mixed-effect on the diagnostic test (\code{\link{lme2}}), a nonparametric bootstrap procedure for clustered data will be applied to obtain the asymptotic covariance matrix the estimated covariate-specific optimal thresholds in original scale (see To et al. 2021, for more details).

The \code{control} argument is a list that can supply any of the following components:
\tabular{ll}{
\code{method.optim} \tab \cr
\tab Name of optimization method to be used. There are three options: \code{"L-BFGS-B"}, \code{"BFGS"} and \code{"Nelder-Mead"}. \cr
\code{start} \tab \cr
\tab Initial values for the thresholds to be optimized over. If it is \code{NULL}, a starting point will be automatically obtained. \cr
\code{maxit} \tab \cr
\tab The maximum number of iterations. Defaults to 200. \cr
\code{lower, upper} \tab \cr
\tab Bounds on the variables for the "L-BFGS-B" method. Defaults are \code{-Inf} and \code{Inf}. \cr
\code{nR} \tab \cr
\tab The number of bootstrap replicates for estimating the covariance matrix under Box-Cox transformation. Defaults to 250. \cr
\code{parallel} \tab \cr
\tab A logical value. If TRUE, a parallel computing is employed in the bootstrap resampling process. \cr
\code{ncpus} \tab \cr
\tab Number of processes to be used in parallel computing. Default is 2.
}
}
\examples{

data(data_3class)
## One covariate
out1 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1"), name.clust = "id_Clus",
             data = data_3class)

### Estimate covariate-specific optimal thresholds at multiple values of one covariate,
### with 3 methods
out_thres_1 <- optThres3(method = c("GYI", "MV", "CtP"), out_lme2 = out1, x.val = 1,
                         apVar = TRUE)
print(out_thres_1)
plot(out_thres_1)

## Two covariates
out2 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1", "X2"),
             name.clust = "id_Clus", data = data_3class)

### Estimate covariate-specific optimal thresholds at one point, with 3 methods
out_thres_2 <- optThres3(method = c("GYI", "MV", "CtP"), out_lme2 = out2, x.val = c(1, 0),
                         apVar = TRUE)
print(out_thres_2)
plot(out_thres_2)

### Estimate covariate-specific optimal thresholds at three points, with 3 methods
out_thres_3 <- optThres3(method = c("GYI", "MV", "CtP"), out_lme2 = out2,
                         x.val = rbind(c(-0.5, 0), c(0.5, 0), c(0.5, 1)), apVar = TRUE)
print(out_thres_3)
plot(out_thres_3, colors = c("forestgreen", "blue"))

}
\references{
To, D-K., Adimari, G., Chiogna, M. and Risso, D. (2021)
``ROC estimation and threshold selection criteria in three-class classification problems for clustered data''. \emph{Submitted}.
}
