% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lme2.R
\name{plot.lme2}
\alias{plot.lme2}
\title{Plot an lme2 object.}
\usage{
\method{plot}{lme2}(x, file.name = NULL, ...)
}
\arguments{
\item{x}{an object of class "lme2", a result of a call to \code{\link{lme2}}.}

\item{file.name}{File name to create on disk.}

\item{...}{further arguments passed to \code{\link{ggexport}} method, for example, \code{width}, \code{height}.}
}
\description{
Diagnostic plots for the linear mixed-effects fit are obtained by lme2.
}
\details{
\code{plot.lme2} provides three diagnostic plots: Q-Q plots of residuals, Fitted vs. Residuals, and Q-Q plot of cluster effects, based on \code{ggplot()}.
}
\seealso{
\code{\link{lme2}}
}
