% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{print.AUC}
\alias{print.AUC}
\title{Print summary results of AUC}
\usage{
\method{print}{AUC}(
  x,
  digits = max(3, getOption("digits") - 2),
  dig.tst = max(1, min(5, digits - 1)),
  ...
)
}
\arguments{
\item{x}{an object of class "AUC", a result of a call to \code{\link{AUC}}.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.AUC} prints the results for the output of function \code{\link{AUC}}.
}
\details{
\code{print.auc} shows a nice format of the summary table for covariate-specific AUC estimates.
}
\seealso{
\code{\link{AUC}}
}
