% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_vus.R
\name{print.ci_clus_vus}
\alias{print.ci_clus_vus}
\title{Print summary results from ci_clus_vus}
\usage{
\method{print}{ci_clus_vus}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class "ci_clus_vus", a result of \code{\link{ci_clus_vus}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\value{
\code{print.ci_clus_vus} shows a summary table for confidence intervals for covariate-specific VUS.
}
\description{
\code{print.ci_vus} displays the results of the output from \code{\link{ci_clus_vus}}.
}
\details{
\code{print.ci_clus_vus} shows a summary table for confidence interval limits for covariate-specific VUS.
}
\seealso{
\code{\link{clus_vus}}
}
