% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vus.R
\name{print.ci_VUS}
\alias{print.ci_VUS}
\title{Print summary results of ci_VUS}
\usage{
\method{print}{ci_VUS}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class "ci_VUS", a result of a call to \code{\link{ci_VUS}}.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.ci_VUS} prints the results for the output of function \code{\link{ci_VUS}}.
}
\details{
\code{print.ci_VUS} shows a nice format of the summary table for CI for covariate-specific VUS estimates.
}
\seealso{
\code{\link{VUS}}
}
