% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_lme.R
\name{print.clus_lme}
\alias{print.clus_lme}
\title{Print summary results of an clus_lme object}
\usage{
\method{print}{clus_lme}(x, digits = max(3L, getOption("digits") - 3L), call = TRUE, ...)
}
\arguments{
\item{x}{an object of class "clus_lme", a result of \code{\link{clus_lme}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{call}{logical. If set to \code{TRUE}, the matched call will be printed.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\value{
\code{print.clus_lme} returns a summary table for the estimated parameters in the cluster-effect model.
}
\description{
\code{print.clus_lme} displays results of the output from \code{\link{clus_lme}}.
}
\details{
\code{print.clus_lme} shows a summary table for the estimated parameters in the cluster-effect model (continuous diagnostic test in three-class setting).
}
\seealso{
\code{\link{clus_lme}}
}
