% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_opt_thresh_3class.R
\name{print.clus_opt_thres3}
\alias{print.clus_opt_thres3}
\title{Print summary results from \code{clus_opt_thres3}}
\usage{
\method{print}{clus_opt_thres3}(x, digits = 3, call = TRUE, ...)
}
\arguments{
\item{x}{an object of class "clus_opt_thres3", a result of \code{\link{clus_opt_thres3}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{call}{logical. If set to \code{TRUE}, the matched call will be printed.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\value{
\code{print.clus_opt_thres3} returns a summary table for results of covariate-specific optimal pair of thresholds estimation.
}
\description{
\code{print.clus_opt_thres3} displays the results of the output from \code{\link{clus_opt_thres3}}.
}
\details{
\code{print.clus_opt_thres3} shows a summary table for covariate-specific optimal pair of thresholds estimates.
}
\seealso{
\code{\link{clus_opt_thres3}}
}
