% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus_tcfs.R
\name{print.clus_tcfs}
\alias{print.clus_tcfs}
\title{Print summary results from clus_tcfs}
\usage{
\method{print}{clus_tcfs}(x, digits = 3, call = TRUE, ...)
}
\arguments{
\item{x}{an object of class "clus_tcfs", a result of \code{\link{clus_tcfs}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{call}{logical. If set to \code{TRUE}, the matched call will be printed.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\value{
\code{print.clus_tcfs} returns a summary table for covariate-specific TCFs estimates.
}
\description{
\code{print.clus_tcfs} displays the results of the output from \code{\link{clus_tcfs}}.
}
\details{
\code{print.clus_tcfs} shows a summary table for covariate-specific TCFs estimates.
}
\seealso{
\code{\link{clus_tcfs}}
}
