% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lme2.R
\name{print.lme2}
\alias{print.lme2}
\title{Print summary results of an lme2 object}
\usage{
\method{print}{lme2}(x, digits = max(3L, getOption("digits") - 3L), call = TRUE, ...)
}
\arguments{
\item{x}{an object of class "lme2", a result of \code{\link{lme2}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{call}{logical. If \code{TRUE}, the matched call will be printed.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.lme2} displays results of the output from \code{\link{lme2}}.
}
\details{
\code{print.lme2} shows a summary table for the estimated parameters in the cluster-effect model (continuous diagnostic test in three-class setting).
}
\seealso{
\code{\link{lme2}}
}
