% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_3class.R
\name{print.optThres3}
\alias{print.optThres3}
\title{Print summary results from \code{optThres3}}
\usage{
\method{print}{optThres3}(x, digits = 3, call = TRUE, ...)
}
\arguments{
\item{x}{an object of class "optThres3", a result of \code{\link{optThres3}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{call}{logical. If \code{TRUE}, the matched call will be printed.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.optThres3} displays the results of the output from \code{\link{optThres3}}.
}
\details{
\code{print.optThres3} shows a summary table for covariate-specific optimal pair of thresholds estimates.
}
\seealso{
\code{\link{optThres3}}
}
