% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCFs.R
\name{print.TCFs}
\alias{print.TCFs}
\title{Print summary results from TCFs}
\usage{
\method{print}{TCFs}(x, digits = 3, call = TRUE, ...)
}
\arguments{
\item{x}{an object of class "TCFs", a result of \code{\link{TCFs}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{call}{logical. If \code{TRUE}, the matched call will be printed.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.TCFs} displays the results of the output from \code{\link{TCFs}}.
}
\details{
\code{print.TCFs} shows a summary table for covariate-specific TCFs estimates.
}
\seealso{
\code{\link{TCFs}}
}
