% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vus.R
\name{print.VUS}
\alias{print.VUS}
\title{Print summary results of VUS}
\usage{
\method{print}{VUS}(
  x,
  digits = max(3, getOption("digits") - 2),
  dig.tst = max(1, min(5, digits - 1)),
  ...
)
}
\arguments{
\item{x}{an object of class "VUS", a result of a call to \code{\link{VUS}}.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.VUS} prints the results for the output of function \code{\link{VUS}}.
}
\details{
\code{print.vus} shows a nice format of the summary table for covariate-specific VUS estimates.
}
\seealso{
\code{\link{VUS}}
}
