% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC_surface.R
\name{ROCsurface}
\alias{ROCsurface}
\title{Plot the covariate-specific ROC surface for clustered data after fitting the cluster-effect model.}
\usage{
ROCsurface(
  out_lme2,
  x.val,
  step.tcf = 0.01,
  plot = TRUE,
  main = NULL,
  file.name = NULL,
  ellips = FALSE,
  thresholds = NULL,
  ci.level = ifelse(ellips, 0.95, NULL)
)
}
\arguments{
\item{name.test}{name of variable indicating diagnostic test of biomarker in data. It can be also the transformation, for example, \code{"log(y)"}, where the term \code{log} is for the log-transformation, and \code{y} is the name of test.}

\item{name.class}{name of variable indicating disease classes (diagnostic groups) in data.}

\item{name.covars}{an vector of names of covariates containing in data. The vector can contain also the transformation, for example, \code{c("x1", "sqrt(x2)", "I(x3^2)")}.}

\item{name.clust}{name of variable indicating clusters in data.}

\item{data}{a data frame containing the variables in the model.}

\item{levl.class}{an vector of the unique values (as character strings) that (disease) class might have taken, sorted into increasing order of means of test results corresponding to the disease classes (diagnostic groups). If \code{levl.class = NULL}, the levels will be automatically determined based on data, and sorted into increasing order of means of test results corresponding to the disease classes (diagnostic groups).}

\item{boxcox}{a logical value. Default = \code{FALSE}. If set to \code{TRUE}, a Box-Cox transformation will be applied to the model to guarantee the normally assumptions.}

\item{apVar}{a logical value. Default = \code{TRUE}. If set to \code{TRUE}, the covariance matrix for all estimated parameters in model with be obtained by using the sandwich formula.}

\item{interval_lambda}{a vector containing the end-points of the interval to be searched for the Box-Cox parameter, \code{lambda}. Default = (-2, 2).}

\item{...}{additional arguments for \code{\link{lme}}, such as \code{control}, \code{contrasts}.}
}
\value{
\code{lme2} returns an object of class inheriting from "lme2" class.

The function \code{\link{print.lme2}} can be used to print a summary of the results.

An object of class "lme2" is a list containing at least the following components:

\item{call}{the matched call.}
\item{est_para}{the estimate of all parameters in model.}
\item{se_para}{the standard error, obtained by using the sandwich formula.}
\item{vcov_sand}{the estimated covariance matrix for all estimated parameters, obtained by the sandwich formula.}
\item{residual}{a list of the residuals}
\item{fitted}{a list of the fitted values.}
\item{randf}{a vector of the estimated random effects.}
\item{n_coef}{total numbers of coefficients included in model.}
\item{icc}{a estimate of intra-class correlation - ICC}
\item{boxcox}{logical value indicating whether the Box-Cox transformation was implemented or not.}
}
\description{
\code{ROCsurface} estimates and makes a 3D plot of covariate-specific ROC surface for clustered data.
}
\details{
.....
}
\examples{

}
