% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCFs.R
\name{TCFs}
\alias{TCFs}
\title{Estimate the covariate-specific TCFs for clustered data.}
\usage{
TCFs(out_lme2, x.val, thresholds, apVar = FALSE)
}
\arguments{
\item{out_lme2}{an object of class "lme2", a result of a call to \code{\link{lme2}}.}

\item{x.val}{specific value(s) of covariate(s) where the TCFs are computed. In case non-covariate, no value is needed to specify. In case of one covariate, \code{x.val} should be a number or a vector. In case of \eqn{p} covariates (\eqn{p > 1}), \code{x.val} should be a vector containing \eqn{p} values of the covariates; or a matrix with \eqn{p} columns and \eqn{m} rows containing values of the covariates if the user wants to compute TCFs at \eqn{m} points.}

\item{thresholds}{a specified pair of thresholds.}

\item{apVar}{a logical value. If set to \code{TRUE}, the variance-covariance matrix of covariate-specific TCFs is computed.}
}
\value{
\code{TCFs} returns an object of class inheriting from "TCFs" class, which is a list containing at least the following components:

\item{call}{the matched call.}
\item{tcfs_est}{a vector or matrix containing the estimated TCFs.}
\item{tcf_vcov}{a matrix or list of matrices containing the estimated variance-covariance matrices.}
\item{thresholds}{the specified pair of thresholds.}
\item{x.val}{value(s) of covariate(s).}
\item{n_p}{total numbers of the regressors in the model.}

Generic functions such as \code{print} has methods to show the results.
}
\description{
\code{TCFs} estimates covariate-specific True Class Fractions (TCFs) at a specified pair of thresholds of a continuous diagnostic test in a clustered design when subjects can be diagnosed in three ordinal groups. This function allows to compute covariate-specific TCFs at multiple points of covariates.
}
\details{
This function implements estimation method in To et al. (2021) for estimating covariate-specific TCFs at a specified pair of thresholds of a continuous diagnostic test in a clustered design when subjects can be diagnosed in three ordinal groups. The estimator is based on the results of fitting the linear mixed-effect model on the diagnostic tests, which is done by using \code{\link{lme2}} with REML approach. The asymptotic variance-covariance matrix of the estimated covariate-specific TCFs is approximated through the Delta method. Notice that, if the Box-Cox transformation was applied for the linear mixed-effect on the diagnostic tests (\code{\link{lme2}}), the input thresholds have to be in the original scale.

Before applying the estimation, a quick check for the monotone ordering assumption will be performed. That is, for given values of covariates, three predicted means of three diagnostic groups will be compared. If the assumption does not meet, the covariate-specific TCFs at the values of covariates will be not estimated.
}
\examples{
data(data_3class)
## One covariate
out1 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1"), name.clust = "id_Clus",
             data = data_3class)

### Estimate TCFs at one single value of X1, (t1, t2) = (1, 4)
out_tcfs_1 <- TCFs(out_lme2 = out1, x.val = 1, thresholds = c(1, 4), apVar = TRUE)
print(out_tcfs_1)

## Two covariates
out2 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1", "X2"),
             name.clust = "id_Clus", data = data_3class)

### Estimate covariate-specific TCFs at point (X1, X2) = (1, 0), and (t1, t2) = (1, 4)
out_tcfs_2 <- TCFs(out_lme2 = out2, x.val = c(1, 0), thresholds = c(1, 4), apVar = TRUE)
print(out_tcfs_2)

### Estimate covariate-specific TCFs at three points and (t1, t2) = (1, 4)
out_tcfs_3 <- TCFs(out_lme2 = out2, x.val = rbind(c(-0.5, 0), c(0.5, 0), c(0.5, 1)),
                   thresholds = c(1, 4), apVar = TRUE)
print(out_tcfs_3)

}
\references{
To, D-K., Adimari, G., Chiogna, M. and Risso, D. (2021)
``ROC estimation and threshold selection criteria in three-class classification problems for clustered data''. \emph{Submitted}.
}
