% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lme2.R
\name{plot.lme2}
\alias{plot.lme2}
\title{Plot an lme2 object.}
\usage{
\method{plot}{lme2}(x, file.name = NULL, ...)
}
\arguments{
\item{x}{an object of class "lme2", i.e., a result of \code{\link{lme2}} call.}

\item{file.name}{File name to create on disk.}

\item{...}{further arguments used with \code{\link{ggexport}} function, for example, \code{width}, \code{height}.}
}
\description{
Diagnostic plots for the linear mixed-effect model, fitted by lme2.
}
\details{
\code{plot.lme2} provides three diagnostic plots: Q-Q plots for residuals, Fitted vs. Residuals values, and Q-Q plot for cluster effects, based on \code{ggplot()}.
}
\seealso{
\code{\link{lme2}}
}
