% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lme2.R
\name{plot.lme2}
\alias{plot.lme2}
\title{Diagnostic plots for model fitted by lme2}
\usage{
\method{plot}{lme2}(x, file.name = NULL)
}
\arguments{
\item{x}{an object of class "lme2", a result of a call to \code{\link{lme2}}.}

\item{file.name}{File name to create on disk.}
}
\description{
\code{plot.lme2} provides diagnostic plots (based on \code{ggplot()}) for the residuals and cluster effects.
}
\details{
\code{plot.lme2} shows three diagnostic plots: Q-Q plots of residuals, Fitted vs. Residuals, and Q-Q plot of cluster effects.
}
\seealso{
\code{\link{lme2}}
}
