% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_3class.R
\name{plot.optThres3}
\alias{plot.optThres3}
\title{Plot confidence regions for covariate-specific optimal thresholds.}
\usage{
\method{plot}{optThres3}(
  x,
  ci.level = 0.95,
  colors = NULL,
  xlims,
  ylims,
  size.point = 0.5,
  size.path = 0.5,
  names.labels,
  file.name = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class "optThres3", a result of a call to \code{\link{optThres3}}.}

\item{ci.level}{a confidence level to be used for constructing the confidence interval; default is 0.95.}

\item{colors}{specification color(s) for confidence regions. If specified, the number of colors needs to be equal the number of covariates}

\item{xlims, ylims}{numeric vectors of length 2, giving the x and y coordinates ranges (the possible range of the first and the second threshold).}

\item{size.point, size.path}{numeric value of length 1, giving the amount by which plotting points and lines.}

\item{names.labels}{a optional character vector giving the label name of covariates.}

\item{file.name}{File name to create on disk.}

\item{...}{further arguments passed to \code{\link{ggexport}} method, for example, \code{width}, \code{height}.}
}
\description{
Plot covariate-specific optimal thresholds and the corresponding confidence regions.
}
\details{
\code{plot.optThres3} provides plots for displaying the confidence regions and point estimates of covariate-specific optimal thresholds. The plots are based on \code{ggplot()}.
}
\seealso{
\code{\link{optThres3}}
}
