% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh_3class.R
\name{print.optThres3}
\alias{print.optThres3}
\title{Print summary results of \code{optThres3}}
\usage{
\method{print}{optThres3}(x, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{an object of class "optThres3", a result of a call to \code{\link{optThres3}}.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.optThres3} prints the results for the output of function \code{\link{optThres3}}.
}
\details{
\code{print.optThres3} shows a nice format of the summary table for covariate-specific optimal thresholds estimates.
}
\seealso{
\code{\link{optThres3}}
}
