% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCFs.R
\name{print.TCFs}
\alias{print.TCFs}
\title{Print summary results of TCFs}
\usage{
\method{print}{TCFs}(x, digits = max(3, getOption("digits") - 2), ...)
}
\arguments{
\item{x}{an object of class "TCFs", a result of a call to \code{\link{TCFs}}.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\description{
\code{print.TCFs} prints the results for the output of function \code{\link{TCFs}}.
}
\details{
\code{print.TCFs} shows a nice format of the summary table for covariate-specific TCFs estimates.
}
\seealso{
\code{\link{TCFs}}
}
