% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vus.R
\name{print.VUS}
\alias{print.VUS}
\title{Print summary results from VUS}
\usage{
\method{print}{VUS}(x, digits = 3, call = TRUE, ...)
}
\arguments{
\item{x}{an object of class "VUS", a result of \code{\link{VUS}} call.}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{call}{logical. If set to \code{TRUE}, the matched call will be printed.}

\item{...}{further arguments passed to \code{\link{print}} method.}
}
\value{
\code{print.VUS} returns a summary table for covariate-specific VUS estimates.
}
\description{
\code{print.VUS} displays the results of the output from \code{\link{VUS}}.
}
\details{
\code{print.VUS} shows a summary table for covariate-specific VUS estimates, containing estimates, standard errors, z-values and p-values for the hypothesis testing \eqn{H_0: VUS = 1/6} versus an alternative \eqn{H_A: VUS > 1/6}.
}
\seealso{
\code{\link{VUS}}
}
