% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC_surface.R
\name{ROCsurface}
\alias{ROCsurface}
\title{Plot the covariate-specific ROC surface for clustered data.}
\usage{
ROCsurface(
  out_lme2,
  x.val,
  step.tcf = 0.01,
  main = NULL,
  file.name = NULL,
  ellips = FALSE,
  thresholds = NULL,
  ci.level = ifelse(ellips, 0.95, NULL)
)
}
\arguments{
\item{out_lme2}{an object of class "lme2", a result of a call to \code{\link{lme2}}.}

\item{x.val}{specific value(s) of covariate(s) where the ROC surface are computed. In case non-covariate, no value is needed to specify. In case of one covariate, \code{x.val} should be a number. In case of \eqn{p} covariates (\eqn{p > 1}), \code{x.val} should be a vector containing \eqn{p} values of the covariates.}

\item{step.tcf}{number: increment of the \eqn{p1 = tcf1} and \eqn{p3 = tcf3}.}

\item{main}{the main title (on top) of plot.}

\item{file.name}{path to file to save the plot.}

\item{ellips}{a logical value. If \code{TRUE}, adds an ellipsoidal confidence region for TCFs (True Class Fractions) at a specified pair of thresholds to current plot of covariate-specific ROC surface.}

\item{thresholds}{a specified pair of thresholds, which used to construct the ellipsoid confidence region for TCFs.}

\item{ci.level}{a confidence level to be used for constructing the confidence interval; default is 0.95.}
}
\value{
\code{ROCsurface} returns a 3D \code{rgl} plot of covariate-specific ROC surface.
}
\description{
\code{ROCsurface} estimates and makes a 3D plot of covariate-specific ROC surface of a continuous diagnostic test in a clustered design when subjects can be diagnosed in three ordinal groups.
}
\details{
This function implements estimation method in To et al. (2021) for estimating covariate-specific ROC surface of a continuous diagnostic test in a clustered design when subjects can be diagnosed in three ordinal groups. The estimator is based on the results of fitting the linear mixed-effect model on the diagnostic tests, which is done by using \code{\link{lme2}} with REML approach.

Before applying the estimation, a quick check for the monotone ordering assumption will be performed. That is, for given values of covariates, three predicted means of three diagnostic groups will be compared. If the assumption does not meet, the covariate-specific ROC surface at the values of covariates will be not estimated.

The ellipsoidal confidence region of TCFs at a given pair of thresholds can be constructed by using a normal approximation and plotted in the ROC surface space. The confidence level (default) is 0.95. Notice that, if the Box-Cox transformation was applied for the linear mixed-effect on the diagnostic tests (\code{\link{lme2}}), the thresholds have to input by the original scale. If the constructed confidence region of TCFs is outside the unit cube, a probit transformation will be automatically applied to obtain a appropriate confidence region which is inside the unit cube (see Bantis et. al., 2017).
}
\examples{
\dontrun{
data(data_3class)
## One covariate
out1 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1"), name.clust = "id_Clus",
             data = data_3class)

### plot only covariate-specific ROC surface
ROCsurface(out_lme2 = out1, x.val = 1)

### plot covariate-specific ROC surface and a 95\% ellipsoidal confidence region for TCFs
ROCsurface(out_lme2 = out1, x.val = 1, ellips = TRUE, thresholds = c(0.9, 3.95))

## Two covariates
out2 <- lme2(name.test = "Y", name.class = "D", name.covars = c("X1", "X2"),
             name.clust = "id_Clus", data = data_3class)

### plot only covariate-specific ROC surface
ROCsurface(out_lme2 = out2, x.val = c(1, 1))

### plot covariate-specific ROC surface and a 95\% ellipsoidal confidence region for TCFs
ROCsurface(out_lme2 = out2, x.val = c(1, 1), ellips = TRUE, thresholds = c(0.9, 3.95))
}

}
\references{
To, D-K., Adimari, G., Chiogna, M. and Risso, D. (2021)
``ROC estimation and threshold selection criteria in three-class classification problems for clustered data''. \emph{Submitted}.

Bantis, L. E., Nakas, C. T., Reiser, B., Myall, D., and Dalrymple-Alford, J. C. (2017).
``Construction of joint confidence regions for the optimal true class fractions of Receiver Operating Characteristic (ROC) surfaces and manifolds''. \emph{Statistical methods in medical research}, \bold{26}, 3, 1429-1442.
}
