% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vus.R
\name{VUS}
\alias{VUS}
\title{Estimation of the covariate-specific VUS for clustered data.}
\usage{
VUS(out_lme2, x.val, apVar = FALSE, subdivisions = 1000, ...)
}
\arguments{
\item{out_lme2}{an object of class "lme2", a result of \code{\link{lme2}} call.}

\item{x.val}{specific value(s) of covariate(s) where the optimal pair of thresholds are estimated. In absence of covariate, no values have to be specified. In case of one covariate, \code{x.val} should be a number. In case of \eqn{p} covariates (\eqn{p > 1}), \code{x.val} should be a vector containing \eqn{p} values; or a matrix with \eqn{p} columns and \eqn{m} rows containing values of the covariates if the user wants to estimate at \eqn{m} points.}

\item{apVar}{logical value. If set to \code{TRUE}, the standard error for (estimated) covariate-specific VUS are estimated.}

\item{subdivisions}{the maximum number of subintervals used to approximate integral. Default is 1000.}

\item{...}{additional arguments to be passed to \code{\link[stats]{integrate}}}
}
\value{
\code{VUS} returns an object of class "VUS" which is a list containing at least the following components:

\item{call}{the matched call.}
\item{vus_est}{a vector containing the estimated covariate-specific VUS.}
\item{vus_se}{a vector containing the standard errors.}
\item{mess_order}{a diagnostic message from checking the monontone ordering.}
\item{x.val}{value(s) of covariate(s).}
\item{n_p}{total number of regressors in the model.}

Generic functions such as \code{print} is also used to show the results.
}
\description{
This function estimates the covariate-specific VUS of a continuous diagnostic test in the setting of clustered data as described in Xiong et al. (2018). This function allows to estimate covariate-specific VUS at multiple points for covariates.
}
\details{
This function implements method in Xiong et al. (2018) for estimating covariate-specific VUS of a continuous diagnostic test in a clustered design with three ordinal groups. The estimator is based on results from \code{\link{lme2}}, which used REML approach. The standard error of the estimated covariate-specific VUS is approximated through the Delta method.

Before performing estimation, a check for the monotone ordering assumption is performed. This mean that, for the fixed values of covariates, three predicted means values for test results in three diagnostic groups are compared. If the assumption does not meet, the covariate-specific VUS at the values of covariates are not estimated. In addition, this function also performs the statistical test, \eqn{H_0: VUS = 1/6} versus an alternative of interest.
}
\examples{
data(data_3class)
## One covariate
out1 <- lme2(fixed.formula = Y ~ X1, name.class = "D", name.clust = "id_Clus",
             data = data_3class)

### Estimate covariate-specific VUS at one value of one covariate
out_vus1 <- VUS(out1, x.val = 0.5, apVar = TRUE)
ci_VUS(out_vus1, ci.level = 0.95)

### Estimate covariate-specific VUS at multiple values of one covariate
out_vus2 <- VUS(out1, x.val = c(-0.5, 0, 0.5), apVar = TRUE)
ci_VUS(out_vus2, ci.level = 0.95)

## Two covariates
out2 <- lme2(fixed.formula = Y ~ X1 + X2, name.class = "D", name.clust = "id_Clus",
             data = data_3class)

### Estimate covariate-specific VUS at one point
out_vus3 <- VUS(out2, x.val = c(1.5, 1), apVar = TRUE)
ci_VUS(out_vus3, ci.level = 0.95)

### Estimate covariate-specific VUS at three points
out_vus4 <- VUS(out2, x.val = rbind(c(-0.5, 0), c(0.5, 0), c(0.5, 1)), apVar = TRUE)
ci_VUS(out_vus4, ci.level = 0.95)

}
\references{
Xiong, C., Luo, J., Chen L., Gao, F., Liu, J., Wang, G., Bateman, R. and Morris, J. C. (2018)
``Estimating diagnostic accuracy for clustered ordinal diagnostic groups in the three-class case -- Application to the early diagnosis of Alzheimer disease''.
\emph{Statistical Methods in Medical Research}, \bold{27}, 3, 701-714.
}
